Completed Merge with dev.
Simplified the code in all controllers by removing redundant code, and code that is not needed.
Removed un-needed controllers, and created the correct controllers

Simplified/cleaned the routes.
Deleted un-needed routes, and created the correct routes.
