--- controllers/explorerController.js
@@ -1,1 +1,0 @@
-//logic related to explorer routes goes here
