+++ views/pages/addQuest.ejs
@@ -13,11 +13,11 @@
 
 <body class="bg-gray-100 p-4 flex flex-col gap-6 text-sm">
     <div >
-        <a href="/researcher/Dashboard" class="text-green-600 font-medium">← Back to Dashboard</a>
+        <a href="/user/researcher" class="text-green-600 font-medium">← Back to Dashboard</a>
     </div>
     
 
-    <form action="/quests/createQuest" method="POST" class="flex flex-col gap-4 bg-white p-4 rounded shadow-md">
+    <form action="/quests/addQuest" method="POST" class="flex flex-col gap-4 bg-white p-4 rounded shadow-md">
         <h1 class="text-xl font-bold text-center">Create a New Quest</h1>
         <div>
             <label for="title" class="font-semibold block mb-1">Title</label>
