+++ views/pages/researcherDashboard.ejs
@@ -54,7 +54,7 @@
                             class="absolute bottom-full bg-black text-white text-xs rounded px-2 py-1 group-hover:opacity-100 mb-2 right-0  opacity-0   mb-2 right-0">
                             Add Quests
                         </div>
-                        <a href="/researcher/add-quest" class="bg-green-700 p-2 rounded-full cursor-pointer block">
+                        <a href="/quests/addQuest" class="bg-green-700 p-2 rounded-full cursor-pointer block">
                             <img src="/images/add.svg" class="h-8 w-8" alt="Add Quest">
                         </a>
                     </div>
